buissness problem
Data provided is related to gold prices. The objective is to understand the underlying structure in your dataset and come up with a suitable forecasting model which can effectively forecast gold prices for next 30 days.
This forecast model will be used by gold exporting and gold importing companies to understand the metal price movements and accordingly set their revenue expectations.